requirements.txt
google-api-python-client==2.154.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
requests==2.32.3
openpyxl==3.1.5
google-genai==0.7.0
